// src/__generated__/scheduleDevicePickerRelay_schedule.graphql.ts
export type scheduleDevicePickerRelay_schedule$data = {
  readonly assignedDevices: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly machineId: string;
        readonly " $fragmentSpreads": FragmentRefs<"deviceSelectorFields_machine">;
      };
    }>;
    readonly filteredCount: number;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
    };
  };
  readonly id: string;
  readonly " $fragmentType": "scheduleDevicePickerRelay_schedule";
};
export type scheduleDevicePickerRelay_schedule$key = {
  readonly " $data"?: scheduleDevicePickerRelay_schedule$data;
  readonly " $fragmentSpreads": FragmentRefs<"scheduleDevicePickerRelay_schedule">;
};